Issue-1 Support for drive / directory selection
Updated flow for files extension, user is provided with list of drives / top level directories in their system. On selection of drive the directory hierarchy is generated from the java backend. User is presented with a browsable tree view that allows user to select 1 or more directories to get file list.